requirements.txt
django
djangorestframework
django-cors-headers
python-dotenv
coreapi
python-dotenv
pandas